% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export.freqshft}
\alias{export.freqshft}
\title{Save the shifted frequencies}
\usage{
export.freqshft(frequency_shifts, output_file)
}
\arguments{
\item{frequency_shifts}{table of allele frequency shifts, as output by freqShifter()}

\item{output_file}{path to savefile}
}
\value{
nothing
}
\description{
Save the shifted frequencies
}
\examples{
\donttest{
merged_frequencies.filename <- system.file("extdata",
"merged_frequencies.example_data.tbl", package = "PopPsiSeqR")
frequencies.bg <- import.freqtbl(merged_frequencies.filename)
frequency_shifts.bg <- freqShifter(frequencies.bg)
export.freqshft(frequency_shifts.bg , tempfile())
}
}
