% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopPsiSeqR-package.R
\docType{package}
\name{PopPsiSeqR-package}
\alias{PopPsiSeqR}
\alias{PopPsiSeqR-package}
\title{PopPsiSeqR: Process and Visualize Evolve & Resequence Experiments}
\description{
Handle data from evolve and resequence experiments. Measured allele frequencies (e.g., from variants called from high-throughput sequencing data) are compared using an update of the PsiSeq algorithm (Earley, Eric and Corbin Jones (2011) \doi{10.1534/genetics.111.129445}). Functions for saving and loading important files are also included, as well as functions for basic data visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csoeder/PopPsiSeq}
  \item \url{https://github.com/csoeder/PopPsiSeqR}
  \item Report bugs at \url{https://github.com/csoeder/PopPsiSeqR/issues}
}

}
\author{
\strong{Maintainer}: Charles Soeder \email{csoeder@email.unc.edu} [copyright holder]

}
\keyword{internal}
