% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differ.R
\name{subTractor}
\alias{subTractor}
\title{Subtractor}
\usage{
subTractor(
  data_one,
  data_two,
  treament_name = "pseudoparent",
  field = "avg_simward_AFshift",
  hoarder = FALSE
)
}
\arguments{
\item{data_one}{first bedfile}

\item{data_two}{second bedfile}

\item{treament_name}{what column is the independent variable}

\item{field}{what column is the variable being compared}

\item{hoarder}{whether or not to retain the other data}
}
\value{
GRanges of the difference
}
\description{
Subtractor
}
\examples{
\donttest{ lab_sechellia.filename <- system.file("extdata",
"wild_sechellia.example_data.bed", package = "PopPsiSeqR")
lab.bg <- import.smvshift(lab_sechellia.filename)
lab.bg$sechellia <- "lab"
wild_sechellia.filename <- system.file("extdata",
"lab_sechellia.example_data.bed", package = "PopPsiSeqR")
wild.bg <- import.smvshift(wild_sechellia.filename)
wild.bg$sechellia <- "wild"
sub.traction <- subTractor(lab.bg, wild.bg ,treament_name = "sechellia")
}
}
