% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export.freqshft}
\alias{export.freqshft}
\title{Save the shifted frequencies}
\usage{
export.freqshft(frequency_shifts, output_file)
}
\arguments{
\item{frequency_shifts}{table of allele frequency shifts, as output by freqShifter()}

\item{output_file}{path to savefile}
}
\description{
Save the shifted frequencies
}
