% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopPsiSeqR-package.R
\docType{package}
\name{PopPsiSeqR-package}
\alias{PopPsiSeqR}
\alias{PopPsiSeqR-package}
\title{PopPsiSeqR: Process and Visualize Evolve & Resequence Experiments}
\description{
Handles data from evolve and resequence experiments. Measured allele frequencies (eg, from variants called from hight-throughput sequencing data) are compared using an update of the PsiSeq algorithm. Functions for saving and loading important files are also included, as well as functions for basic data visiualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csoeder/PopPsiSeq}
  \item \url{https://github.com/csoeder/PopPsiSeqR}
}

}
\author{
\strong{Maintainer}: Charles Soeder \email{csoeder@email.unc.edu}

}
\keyword{internal}
