% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differ.R
\name{subTractor}
\alias{subTractor}
\title{Subtractor}
\usage{
subTractor(
  data_one,
  data_two,
  treament_name = "pseudoparent",
  field = "avg_simward_AFshift",
  hoarder = FALSE
)
}
\arguments{
\item{data_one}{first bedfile}

\item{data_two}{second bedfile}

\item{treament_name}{what column is the independent variable}

\item{field}{what column is the variable being compared}

\item{hoarder}{whether or not to retain the other data}
}
\value{
GRanges of the difference
}
\description{
Subtractor
}
